Many changes related to passing filename information from the part to the modellist. The Kompare::Infostruct was already in place to update everyone about changes but it was not used to communicate between the part and model this commit fixes that.
The real reason for this commit is that it fixes saving remote files. 
Do not give a local URL and pretend it is a remote URL, does not work. 
KIO::NetAccess::upload then simply does nothing because the file 
already exists.  

Refresh is fixed now. Refetch the files and dont reuse the local files
to determine the differences. Now that saving remote works it showed the
old result again. Some had already seen this, now it was easy to figure
out where it came from.

Fixed a silly copy and paste error where for the destination directory
source was used. Always fun these bugs.

Updated the caption in the titlebar to now have a " -- " to separate the
files so it is more clear which files/directories are compared. I read
somewhere in a bugreport people did not like this so fixed as well.

CCBUG:76904
CCBUG:81088
CCBUG:126870 Patch applied, second hunk not since this is not related

backport revision 930684 by bruggie from trunk

svn path=/branches/KDE/4.2/kdesdk/kompare/; revision=930700